+++ resources/views/crud/formfields/int.blade.php
@@ -7,8 +7,11 @@
         class="form-control"
         name="{{ $fieldName }}"
         type="number"
-        placeholder="{{ $prop['rules']['min'] }}"
-        value="{{ old($fieldName) ?? $model->{$fieldName} ?? $prop['rules']['min'] }}"
+        placeholder="{{ $prop['placeholder'] }}"
+        @if(isset($unsigned))
+        min="0"
+        @endif
+        value="{{ old($fieldName) ?? $model->{$fieldName} ?? $prop['rules']['min'] ?? 0 }}"
         @if(isset($prop['rules']['max']))max="{{ $prop['rules']['max'] }}"@endif
         @if(isset($prop['rules']['min']))min="{{ $prop['rules']['min'] }}"@endif
         @if(isset($prop['rules']['step']))step="{{ $prop['rules']['step'] }}"@endif