+++ src/Enums/CrudType.php
@@ -17,6 +17,8 @@ use Kwaadpepper\Enum\BaseEnum;
  * @method static self enum()
  * @method static self order()
  * @method static self color()
+ * @method static self unsignedint()
+ * @method static self unsignedfloat()
  * @method static self belongsTo()
  * @method static self belongsToMany()
  */
@@ -38,6 +40,8 @@ class CrudType extends BaseEnum
             'enum' => 9,
             'order' => 10,
             'color' => 11,
+            'unsignedint' => 12,
+            'unsignedfloat' => 13,
             'belongsTo' => 21,
             'belongsToMany' => 22
         ];
@@ -58,6 +62,8 @@ class CrudType extends BaseEnum
             'enum' => trans('crud-policies::crud.enum'),
             'order' => trans('crud-policies::crud.order'),
             'color' => trans('crud-policies::crud.color'),
+            'unsignedint' => trans('crud-policies::crud.unsignedint'),
+            'unsignedfloat' => trans('crud-policies::crud.unsignedfloat'),
             'belongsTo' => trans('crud-policies::crud.belongsTo'),
             'belongsToMany' => trans('crud-policies::crud.belongsToMany')
         ];